Update README.md
Thanks to Allcon Ltd for the inspiration to start this thing. Right now, there's no real structure in terms of folders etc.

I will put up python scripts etc as I come across or build them. Full disclosure: I am not really a coder, and ChatGPT helps a lot, but 

it's all about using the right tools properly, and this is a fantastic way to leverage AI. 
First item up, a small python script that uses the Gnews API to find ten headlines relating to whatever search terms you enter into a little 

search window. It's ugly and minimal, but it works. Play around with it, Feedback, comments, suggestions for features, feel free to give them!